requirements.txt
# ---- Django Core ----
Django==5.0.2

# ---- Django REST Framework ----
djangorestframework==3.15.1

# ---- JWT Authentication ----
djangorestframework-simplejwt==5.3.1

# ---- Database (optional: PostgreSQL for production) ----
psycopg2-binary==2.9.9

# ---- CORS for React ↔ Django communication ----
django-cors-headers==4.3.1

# ---- Environment variables (secure secrets) ----
python-dotenv==1.0.1
